floating layers, clickable layer menu, scrollable layers
Added scrollable layers, made the whole page fixed.
So, now the user can simply add layers without hassle.
Clickable layer menu: Now user can use a custom button to hide
or show the layers menu to add to the architecture.
Scrollable layers: the whole page is fixed but only the layers are scrollable with scroll on the left side.